Fix for ELLIPSE not drawing the ellipse properly.
The ELLIPSE code in this BASIC doesn't support the rotation, but it
does draw the ellipse in the wrong place - the top most point that
the ellipse is meant to hit should be above the centre point. However,
it was being drawn at an angle because the top most point was above
the x axis intercept.

The fix is just to use a negative relative plot position.